feat: set type of api responses

+++ src/api/todoList.ts
@@ -1,22 +1,54 @@
 import axios from 'axios';
-import { todoParams } from '../typings/todoList';
+import {
+  getTodoResponse,
+  todoParams,
+  createTodoResponse,
+  deleteTodoResponse,
+  getTodoByIdResponse,
+  updateTodoResponse,
+} from '../typings/todoList';
 
-export const getTodo = () => {
+const checkAuthToken = () => {
+  const authToken = localStorage.getItem('token');
+  return authToken ? true : false;
+};
+
+export const getTodo = (): Promise<getTodoResponse> | undefined => {
+  if (!checkAuthToken()) {
+    return;
+  }
+
   return axios.get('/todos');
 };
 
-export const getTodoById = (id: string) => {
+export const getTodoById = (id: string): Promise<getTodoByIdResponse> | undefined => {
+  if (!checkAuthToken()) {
+    return;
+  }
+
   return axios.get(`todos/${id}`);
 };
 
-export const createTodo = (todo: todoParams) => {
+export const createTodo = (todo: todoParams): Promise<createTodoResponse> | undefined => {
+  if (!checkAuthToken()) {
+    return;
+  }
+
   return axios.post('/todos', todo);
 };
 
-export const updateTodoList = (id: string, todo: todoParams) => {
+export const updateTodoList = (id: string, todo: todoParams): Promise<updateTodoResponse> | undefined => {
+  if (!checkAuthToken()) {
+    return;
+  }
+
   return axios.put(`/todos/${id}`, todo);
 };
 
-export const deleteTodoList = (id: string) => {
+export const deleteTodoList = (id: string): Promise<deleteTodoResponse> | undefined => {
+  if (!checkAuthToken()) {
+    return;
+  }
+
   return axios.delete(`/todos/${id}`);
 };
